# Fill GarbageByScore, GroupByGarbageBlockCount once when compaction map is loaded