Benchmark results for disk with size 100_TB
before changes:
| Benchmark | Time | CPU |
|--------|--------|--------|
| BenchmarkUpdate_100_TB | 1.3232e+10 ns | 1.3232e+10 ns |

after changes:
| Benchmark | Time | CPU |
|--------|--------|--------|
| BenchmarkUpdate_100_TB | 7694792291 ns | 7694446694 ns |

As next step need to think about asynchronous loading compaction map.